Fixed Yehuda Moon to work on non-existant dates
Yehuda Moon and the Kickstand Cyclery was failing before on future dates: it
was relying on a title to have a version of pub_date in the drop-down comic
selection list.  This makes sense:  If you ask for date X, you expect to get
date X's comic or no comic.  That's not how they do it: they just give you the
most recent one if it doesn't exist.

So, there were two bugs:  relying on pub_date's comic to exist as passed, and
not trapping a possible error in getting the title gracefully.  They were both
fixed here.