Added GoComics helper and optional REQ headers
To compliment the comics.com parser, I wrote the gocomics.com parser.  There
were a few comics that were hacking in some weird value into a ucomics image
(Calvin and Hobbes, Foxtrot, etc); I didn't know how to extend this.  It
probably makes more sense to have an "actual" loader that has more generic
abilities for a site that now has so many comics.

As part of that, GoComics.com does User-Agent checks to make sure you're a real
browser (or at least know how to fake one).  So, I also passed headers into the
underlying urllib2 call inside lxmlparser.py.  You can now specify any headers
that you want to pass into the request by supplying the 'headers' file.

This still has to get documented; that's coming in the next commit.  Then a
whole lot of comics that are going to rely on this method that are hosted on
GoComics.com.