Fixed a bug in HEAD^^'s changes; added comics
The bug was that I originally named the header 'referer' until I realized that
it wasn't going to scale in case we had to add other headers later.  I don't
know if Stein wants to add some interface to the headers as a nice procelain on
the raw dictionary interface, but I'm assuming that functionality over a 1:1
header-function explosion is OK here as long as it's well-documented.

So, the referer is gone and the headers is here.  It will be quite useful if we
ever have to have things like Referer: or Accepted-Encodings: (say some site
forces gzip unless you turn it off or something).

Oh, with this functionality, I was able to rapid-fire add comics.  I have:

For Better or For Worse
Non Sequitur
Pluggers
Tank McNamara
The Boondocks