Added a link to the official site on "no data"
Whenever I am pulling comics and there's no release in the time range, I'm
wondering if the crawler broke or couldn't find some data for that day.  The
next thing I want to do is go check the official site to see if there's a comic
there for the date I'm looking at.  If so, I want to debug the crawler and
submit patches upstream.

Putting this link on the page (as well as the commit in HEAD^ to hyperlink the
printed on success in case I think it's old) should yield faster debugging of
crawlers that aren't pulling all the results that we think they should.