Migrated virtualenv / debug settings to local
Before, you had to have a static virtualenv address and comment / uncomment the
address in deploy.wsgi.  This is kind of weird and not like the rest of the
program, so I migrated those settings to the local Django setup configuration
file and then put logic in deploy.wsgi to handle all of that in a sane way.

It adds complexity to deploy.wsgi to the benefit of simplifying local.py and
what one would have to do to get a comics installation running.